Correct the explanation of this gem thanks to feedback from sporkmonger
The README made it sound like there was something incorrect about the
way Addressable parsed scp-style git repository locations, when in fact
scp-style uris are not intended to be parsed by Addressable (because
they are not URIs). Gitable is intended to support git repository
locations only, and not be a general solution for URI parsing.